// src/mock/common/types/wardStats.ts
export interface WardStats {
    wardId: string;

    academicYear: string;

    schoolCount: number;

    campusCount: number;

    teacherCount: number;

    studentCount: number;

    twoSessionCount: number;

    boardingCount: number;

    breakfastCount: number;

    lunchCount: number;

    policyStudentCount: number;

    enrollmentChangeCount: number;
}